Align contract imports and move home-location helpers
