+++ src/presenters/brief-json/render-brief-json.test.ts
@@ -1,8 +1,8 @@
 import { describe, expect, it } from 'vitest';
 import { renderBriefAsJson } from './render-brief-json.js';
-import { BRIEF_JSON_SCHEMA_VERSION } from '../../types/brief.js';
+import { BRIEF_JSON_SCHEMA_VERSION } from '../../contracts/index.js';
 import type { EditorialDecision } from '../../app/run-photo-brief/contracts.js';
-import type { ScoredForecastContext } from '../../types/scored-forecast.js';
+import type { ScoredForecastContext } from '../../contracts/index.js';
 
 function makeScoredContext(): ScoredForecastContext {
   return {
