+++ src/presenters/site/sections/long-range.ts
@@ -1,6 +1,6 @@
 import { esc } from '../../../lib/utils.js';
 import { clockToMinutes, minutesToClock } from '../../../domain/windowing/index.js';
-import type { DarkSkyAlertCard, LongRangeCard, RunTimeContext } from '../../../types/brief.js';
+import type { DarkSkyAlertCard, LongRangeCard, RunTimeContext } from '../../../contracts/index.js';
 import { C } from '../../shared/brief-primitives.js';
 import { sCard, sChip, sScorePill, sPill } from './shared.js';
 import { scoreState } from '../../shared/brief-primitives.js';
