+++ src/presenters/site/sections/spur-of-moment.ts
@@ -1,6 +1,6 @@
 import { esc } from '../../../lib/utils.js';
 import { displayTag } from '../../shared/window-helpers.js';
-import type { SpurOfTheMomentSuggestion } from '../../../types/brief.js';
+import type { SpurOfTheMomentSuggestion } from '../../../contracts/index.js';
 import { sCard, sChip } from './shared.js';
 import { C } from '../../shared/brief-primitives.js';
